Change report api to server action
