+++ lib/actions.ts
@@ -4,6 +4,7 @@ import { headers } from "next/headers";
 import { redirect } from "next/navigation";
 import sharp from "sharp";
 import { LoginUser, type SignUpType, User } from "./schemas";
+import { getUserId } from "./getUserId";
 
 export const createPostAction = async (
 	formData: FormData,
@@ -108,4 +109,18 @@ export const signUpAction = async (
 	});
 	if (error) return { status: "error" };
 	return { status: "ok" };
+};
+export const reportPostAction = async (
+	formData: FormData,
+	post_id: string,
+): Promise<{ status: "error" | "ok" }> => {
+	const supabase = await createClient();
+	const id = await getUserId();
+	const description = formData.get("description");
+	const type = formData.get("type");
+	const { error } = await supabase
+		.from("report")
+		.insert({ description, type, post_id, id });
+
+	return error?.message ? { status: "error" } : { status: "ok" };
 };
